editoast: move interpolate_arrival_time() from track_occupancy to projection
